move redux and services into a common folder
