+++ src/common/redux/appshell/reducers.js
