+++ src/common/redux/auto-update/reducers.js
