+++ src/common/redux/blocks/reducers.js
