+++ src/common/redux/middleware/analytics/index.js
