+++ src/common/services/ChainService.js
@@ -20,7 +20,7 @@ class ChainService extends EventEmitter {
   }
 
   start() {
-    let chainPath = path.join(__dirname, "../chain/", "chain.js")
+    let chainPath = path.join(__dirname, "../../chain/", "chain.js")
     const options = {
       stdio: [ 'pipe', 'pipe', 'pipe', 'ipc' ]
     };
