+++ src/common/services/GoogleAnalyticsService.js
@@ -1,5 +1,5 @@
 import ua from 'universal-analytics'
-import * as pkg from '../../package.json'
+import * as pkg from '../../../package.json'
 
 const ganacheAnalyticsId = 'UA-83874933-5'
 const ganacheUrl = 'http://truffleframework.com/ganache'
