+++ src/renderer/init/Logs.js
@@ -3,7 +3,7 @@ import { ipcRenderer } from 'electron'
 import {
   ADD_LOG_LINES,
   addLogLines
-} from '../../redux/logs/actions'
+} from '../../common/redux/logs/actions'
 
 export function initLogs(store) {
   ipcRenderer.on(ADD_LOG_LINES, (event, lines) => {
