+++ src/renderer/init/Network.js
@@ -3,7 +3,7 @@ import { ipcRenderer } from 'electron'
 import {
   SET_INTERFACES,
   setInterfaces
-} from '../../redux/network/actions'
+} from '../../common/redux/network/actions'
 
 export function initNetwork(store) {
   ipcRenderer.on(SET_INTERFACES, (event, interfaces) => {
