+++ src/renderer/screens/appshell/TopNavbar.js
@@ -2,11 +2,11 @@ import React, { Component } from 'react'
 import { Link } from 'react-router'
 
 import connect from '../helpers/connect'
-import * as Search from '../../../redux/search/actions'
-import { setSystemError } from '../../../redux/core/actions'
-import { setUpdateAvailable } from '../../../redux/auto-update/actions'
-import { closeWorkspace, saveWorkspace } from '../../../redux/workspaces/actions'
-import { clearLogLines } from '../../../redux/logs/actions'
+import * as Search from '../../../common/redux/search/actions'
+import { setSystemError } from '../../../common/redux/core/actions'
+import { setUpdateAvailable } from '../../../common/redux/auto-update/actions'
+import { closeWorkspace, saveWorkspace } from '../../../common/redux/workspaces/actions'
+import { clearLogLines } from '../../../common/redux/logs/actions'
  
 import Spinner from '../../components/spinner/Spinner'
 import OnlyIf from '../../components/only-if/OnlyIf'
