+++ src/renderer/screens/blocks/BlockCard.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import connect from '../helpers/connect'
-import * as Blocks from '../../../redux/blocks/actions'
-import * as Transactions from '../../../redux/transactions/actions'
+import * as Blocks from '../../../common/redux/blocks/actions'
+import * as Transactions from '../../../common/redux/transactions/actions'
 
 import Moment from 'react-moment'
