+++ src/renderer/screens/config/ConfigScreen.js
@@ -4,8 +4,8 @@ import cloneDeep from 'lodash.clonedeep'
 import isEqual from 'lodash.isequal'
 import connect from '../helpers/connect'
 
-import * as Core from '../../../redux/core/actions'
-import * as Config from '../../../redux/config/actions'
+import * as Core from '../../../common/redux/core/actions'
+import * as Config from '../../../common/redux/config/actions'
 
 import ServerScreen from './ConfigScreens/ServerScreen'
 import AccountsScreen from './ConfigScreens/AccountsScreen'
