+++ src/renderer/screens/transactions/RecentTransactions.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 
 import connect from '../helpers/connect'
 
-import * as Transactions from '../../../redux/transactions/actions'
+import * as Transactions from '../../../common/redux/transactions/actions'
 
 import TxList from './TxList'
